Fix building with celt in non-standard prefix.
This moves the celt parts of JackNetTool into JackNetCelt so celt
headers are not undeliberately included from JackNetDriver (which is
compiled without celt).
